+++ src/App/Emoji.php
@@ -41,13 +41,17 @@ class Emoji {
 		<?php
 	}
 
+	/**
+	 * @param $lib
+	 *
+	 * @return string
+	 */
 	private function emoji_url() {
-		if ( $this->get_option( 'static_cdn', 'editor_basics' ) === WP_EDITORMD_URL ) {
-			$emoji_url = '//cdnjs.cloudflare.com/ajax/libs/emojify.js/1.1.0';
+		if ( $this->get_option( 'static_cdn', 'editor_basics' ) === '//cdn.jsdelivr.net' ) {
+			$emoji_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/npm/emojify.js@1.1.0/dist';
 		} else {
 			$emoji_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/emojify.js/1.1.0';
 		}
-
 		return $emoji_url;
     }