+++ src/App/Sequence.php
@@ -12,16 +12,52 @@ class Sequence {
 	}
 
 	public function sequence_enqueue_scripts() {
-		wp_enqueue_script( 'jQuery-CDN', $this->get_option( 'static_cdn', 'editor_basics' ) . '/jquery/1.12.4/jquery.min.js', array(), '1.12.4', true );
-		wp_enqueue_script( 'Underscore', $this->get_option( 'static_cdn', 'editor_basics' ) . '/underscore.js/1.9.0/underscore-min.js', array(), '1.9.0', true );
-		wp_enqueue_script( 'Raphael', $this->get_option( 'static_cdn', 'editor_basics' ) . '/raphael/2.2.7/raphael.min.js', array(), '2.2.7', true );
-		wp_enqueue_script( 'Sequence', $this->get_option( 'static_cdn', 'editor_basics' ) . '/js-sequence-diagrams/1.0.6/sequence-diagram-min.js', array(), '1.0.6', true );
+		wp_enqueue_script( 'jQuery-CDN', $this->sequence_url('jquery'), array(), '1.12.4', true );
+		wp_enqueue_script( 'Underscore', $this->sequence_url('underscore'), array(), '1.9.0', true );
+		wp_enqueue_script( 'Raphael', $this->sequence_url('raphael'), array(), '2.2.7', true );
+		wp_enqueue_script( 'Sequence',  $this->sequence_url('sequence'), array(), '1.0.6', true );
 	}
 
 	public function sequence_wp_footer_script() {
 		$seqStyle = $this->get_option( 'sequence_style', 'editor_sequence' );
 		$script = '<script type="text/javascript">(function($){$(document).ready(function(){$(".diagram").sequenceDiagram({theme: "' . $seqStyle . '"})})})(jQuery)</script>';
 		echo $script;
+	}
+
+	private function sequence_url($lib) {
+		if ( $this->get_option( 'static_cdn', 'editor_basics' ) === '//cdn.jsdelivr.net' ) {
+			switch ($lib) {
+				case 'jquery':
+					$lib_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/npm/jquery@3.3.1/dist/jquery.min.js';
+					break;
+				case 'underscore':
+					$lib_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/npm/underscore@1.9.0/underscore-min.js';
+					break;
+				case 'raphael':
+					$lib_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/npm/raphael@2.2.7/raphael.min.js';
+					break;
+				case 'sequence':
+					$lib_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/npm/js-sequence-diagrams@10.0.6/build/sequence-diagram-min.js';
+					break;
+			}
+		} else {
+			switch ($lib) {
+				case 'jquery':
+					$lib_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/jquery/3.3.1/jquery.min.js';
+					break;
+				case 'underscore':
+					$lib_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/underscore.js/1.9.0/underscore-min.js';
+					break;
+				case 'raphael':
+					$lib_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/raphael/2.2.7/raphael.min.js';
+					break;
+				case 'sequence':
+					$lib_url = $this->get_option( 'static_cdn', 'editor_basics' ) . '/js-sequence-diagrams/1.0.6/sequence-diagram-min.js';
+					break;
+			}
+		}
+
+		return $lib_url;
 	}
 
 	/**